重写createCacheKey.
