+++ mybatis-plus-core/src/main/java/com/baomidou/mybatisplus/core/executor/MybatisBaseExecutor.java
@@ -1,0 +1,84 @@
+package com.baomidou.mybatisplus.core.executor;
+
+import com.baomidou.mybatisplus.core.metadata.IPage;
+
+import org.apache.ibatis.cache.CacheKey;
+import org.apache.ibatis.executor.BaseExecutor;
+import org.apache.ibatis.executor.ExecutorException;
+import org.apache.ibatis.mapping.BoundSql;
+import org.apache.ibatis.mapping.MappedStatement;
+import org.apache.ibatis.mapping.ParameterMapping;
+import org.apache.ibatis.mapping.ParameterMode;
+import org.apache.ibatis.reflection.MetaObject;
+import org.apache.ibatis.session.Configuration;
+import org.apache.ibatis.session.RowBounds;
+import org.apache.ibatis.transaction.Transaction;
+import org.apache.ibatis.type.TypeHandlerRegistry;
+
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+
+/**
+ * 重写执行器 {@link org.apache.ibatis.executor.BaseExecutor}
+ *
+ * @author nieqiurong 2019/4/22.
+ */
+public abstract class MybatisBaseExecutor extends BaseExecutor {
+
+    protected MybatisBaseExecutor(Configuration configuration, Transaction transaction) {
+        super(configuration, transaction);
+    }
+
+    @Override
+    public CacheKey createCacheKey(MappedStatement ms, Object parameterObject, RowBounds rowBounds, BoundSql boundSql) {
+        if (super.isClosed()) {
+            throw new ExecutorException("Executor was closed.");
+        }
+        CacheKey cacheKey = new CacheKey();
+        cacheKey.update(ms.getId());
+        Object offset = rowBounds.getOffset();
+        Object limit = rowBounds.getLimit();
+        if (parameterObject instanceof Map) {
+            Map<?, ?> parameterMap = (Map<?, ?>) parameterObject;
+            Optional<? extends Map.Entry<?, ?>> optional = parameterMap.entrySet().stream().filter(entry -> entry.getValue() instanceof IPage).findFirst();
+            if (optional.isPresent()) {
+                IPage<?> page = (IPage) optional.get().getValue();
+                offset = page.getCurrent();
+                limit = page.getTotal();
+            }
+        } else if (parameterObject instanceof IPage) {
+            IPage<?> page = (IPage) parameterObject;
+            offset = page.getCurrent();
+            limit = page.getTotal();
+        }
+        cacheKey.update(offset);
+        cacheKey.update(limit);
+        cacheKey.update(boundSql.getSql());
+        List<ParameterMapping> parameterMappings = boundSql.getParameterMappings();
+        TypeHandlerRegistry typeHandlerRegistry = ms.getConfiguration().getTypeHandlerRegistry();
+        // mimic DefaultParameterHandler logic
+        for (ParameterMapping parameterMapping : parameterMappings) {
+            if (parameterMapping.getMode() != ParameterMode.OUT) {
+                Object value;
+                String propertyName = parameterMapping.getProperty();
+                if (boundSql.hasAdditionalParameter(propertyName)) {
+                    value = boundSql.getAdditionalParameter(propertyName);
+                } else if (parameterObject == null) {
+                    value = null;
+                } else if (typeHandlerRegistry.hasTypeHandler(parameterObject.getClass())) {
+                    value = parameterObject;
+                } else {
+                    MetaObject metaObject = configuration.newMetaObject(parameterObject);
+                    value = metaObject.getValue(propertyName);
+                }
+                cacheKey.update(value);
+            }
+        }
+        if (configuration.getEnvironment() != null) {
+            // issue #176
+            cacheKey.update(configuration.getEnvironment().getId());
+        }
+        return cacheKey;
+    }
+}
